O_Fibonacci

+++ Codeforces/my-solved-problem-in-ICPC-Assiut-University-Community/Sheet #3 (Arrays)/O_Fibonacci.cpp
@@ -1,16 +1,20 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int main() {
+int main()
+{
     int n;
     cin >> n;
-    int a = 0, b = 1, fib;
+    long long a = 0, b = 1, fib;
 
-    if (n == 1) {
+    if (n == 1) 
+    {
         cout << a;
-    } else if (n == 2) {
+    } else if (n == 2) 
+    {
         cout << b;
-    } else {
+    } else 
+    {
         for (int i = 3; i <= n; i++) 
         {
             fib = a + b;
